Create LICENSE
I do not wish to monetize this code and would like the community to use this if they wish, so I'm putting this under the MIT license because this is an educational material and there are already more performant OCR models available for free that people could use if they so desire. But I believe this project code could be useful for other Computer Science students, and I learned a lot along the way.